bugfix regarding i18n (two methods takes locale as id)
Problem: calling edit_user_url and edit_plan_url with i18n enabled causes that the locale is taken as id and therefore id is nil.
It looks like the routes are expecting two params, and ordering it as it comes.

Fix: using a named hash into the params passed