Fix link_helpers breaking links.
As of d84cdefa084550d7b0693f6a5e72edef81ff13e4, helpers would override the
locale with the parameter passed to the method, which, considering that they
usually are not used that way, resulted in the locale being nil. This way, the
locale was lost.
Also, a small mistake yielded the html format always being passed to URL
generators, which in turn broke links to the root.