Remove an apparently unused, and possibly dangerous route.
The potential danger was: The route did arbitrary redirects. It was attempted to
limit them to local URLs, but this didn't prevent me from redirecting to
external servers. The danger was limited since the route was restricted to POST,
which does limit arbitrary use due to CSRF protection (both on our and on
external servers). It probably still could be exploited in conjunction with
other (XSS) vulnerabilities.

The route was apparently obsoleted in November 2013 with the obsoletion (and
later removal) of the only partial it was used in, in commit
90c98f9905d99794bd40a7f3a250cc94559b6f0c.