Fix possible NPE when changing locale on the account import page.
For some reason, shibboleth auth data persists across the several accesses needed for importing a new account, but gets lost as soon as you try to change locale on the user import page.

This was fixed by writing the auth provider (as accessing caused the error) to the session before the redirect (instead of after) and later restoring it from there, if needed.